translator tool/requirements.txt
azure-functions==1.21.3
ibm-watsonx-ai==1.2.7
